star-shaped asteroids

+++ Floater.pde
@@ -1,8 +1,8 @@
 abstract class Floater //Do NOT modify the Floater class! Make changes in the Spaceship class 
 {   
   protected int corners;  //the number of corners, a triangular floater has 3   
-  protected int[] xCorners;   
-  protected int[] yCorners;   
+  protected float[] xCorners;   
+  protected float[] yCorners;   
   protected color myColor;
   protected int myStroke;
   protected double myCenterX, myCenterY; //holds center coordinates
